+++ kernel/cuckarookernel_bak.go
@@ -17,19 +17,20 @@ typedef u32 node_t;
 typedef u64 nonce_t;
 
 
+#define BASE_SIZE 32L
 #define DUCK_SIZE_A 129L
 #define DUCK_SIZE_B 83L
 
-#define DUCK_A_EDGES (DUCK_SIZE_A * 1024L)
+#define DUCK_A_EDGES (DUCK_SIZE_A * BASE_SIZE)
 #define DUCK_A_EDGES_64 (DUCK_A_EDGES * 64L)
 
-#define DUCK_B_EDGES (DUCK_SIZE_B * 1024L)
+#define DUCK_B_EDGES (DUCK_SIZE_B * BASE_SIZE)
 #define DUCK_B_EDGES_64 (DUCK_B_EDGES * 64L)
 
 #define EDGE_BLOCK_SIZE (64)
 #define EDGE_BLOCK_MASK (EDGE_BLOCK_SIZE - 1)
 
-#define EDGEBITS 29
+#define EDGEBITS 24
 // number of edges
 #define NEDGES ((node_t)1 << EDGEBITS)
 // used to mask siphash output
@@ -70,7 +71,6 @@ bool Read2bCounter(__local u32 * ecounters, const int bucket)
 	return (ecounters[word + 4096] & mask) > 0;
 }
 
-__attribute__((reqd_work_group_size(128, 1, 1)))
 __kernel  void FluffySeed2A(const u64 v0i, const u64 v1i, const u64 v2i, const u64 v3i, __global ulong4 * bufferA, __global ulong4 * bufferB, __global u32 * indexes)
 {
 	const int gid = get_global_id(0);
@@ -170,7 +170,6 @@ __kernel  void FluffySeed2A(const u64 v0i, const u64 v1i, const u64 v2i, const u
 
 }
 
-__attribute__((reqd_work_group_size(128, 1, 1)))
 __kernel  void FluffySeed2B(const __global uint2 * source, __global ulong4 * destination1, __global ulong4 * destination2, const __global int * sourceIndexes, __global int * destinationIndexes, int startBlock)
 {
 	const int lid = get_local_id(0);
@@ -258,7 +257,6 @@ __kernel  void FluffySeed2B(const __global uint2 * source, __global ulong4 * des
 	}
 }
 
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel   void FluffyRound1(const __global uint2 * source1, const __global uint2 * source2, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes, const int bktInSize, const int bktOutSize)
 {
 	const int lid = get_local_id(0);
@@ -319,7 +317,6 @@ __kernel   void FluffyRound1(const __global uint2 * source1, const __global uint
 
 }
 
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel   void FluffyRoundN(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
@@ -380,7 +377,6 @@ __kernel   void FluffyRoundN(const __global uint2 * source, __global uint2 * des
 
 }
 
-__attribute__((reqd_work_group_size(64, 1, 1)))
 __kernel   void FluffyRoundN_64(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
@@ -441,7 +437,6 @@ __kernel   void FluffyRoundN_64(const __global uint2 * source, __global uint2 * 
 
 }
 
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel void FluffyTail(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
@@ -461,7 +456,6 @@ __kernel void FluffyTail(const __global uint2 * source, __global uint2 * destina
 	}
 }
 
-__attribute__((reqd_work_group_size(256, 1, 1)))
 __kernel   void FluffyRecovery(const u64 v0i, const u64 v1i, const u64 v2i, const u64 v3i, const __constant u64 * recovery, __global int * indexes)
 {
 	const int gid = get_global_id(0);
@@ -532,7 +526,6 @@ __kernel   void FluffyRecovery(const u64 v0i, const u64 v1i, const u64 v2i, cons
 // ---------------
 #define BKT_OFFSET 255
 #define BKT_STEP 32
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel   void FluffyRoundNO1(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
@@ -593,7 +586,6 @@ __kernel   void FluffyRoundNO1(const __global uint2 * source, __global uint2 * d
 
 }
 
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel   void FluffyRoundNON(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
@@ -654,7 +646,6 @@ __kernel   void FluffyRoundNON(const __global uint2 * source, __global uint2 * d
 
 }
 
-__attribute__((reqd_work_group_size(1024, 1, 1)))
 __kernel void FluffyTailO(const __global uint2 * source, __global uint2 * destination, const __global int * sourceIndexes, __global int * destinationIndexes)
 {
 	const int lid = get_local_id(0);
